add example output
Added the 5 types of example output. It should be fairly self evident,
the filenames are the zone shortname, the dump type, then the data and
time. Doors are all the doors, GroundItem are groundspawns and non
mobile objects like trade skill containers, NPC are all the spawns that
are up, Zone is the current zone's information, and ZonePoint is all the
data about the zonelines in your current zone.