Pathing/Target
Add the ability to record a targets path. Acquire a target, then do
"/takeadump target" or "/takeadump path". It will continue to record
until you lose the target. It will record an initial point, and a new
point every time the heading changes. So if it is walking in a straight
line it wont spam points while walking the line, just the endpoints when
it turns.