added some minor data validation
In each dump routine I added some variable validation on an element
towards the end of the struct. The ho[e is this will allow the plugin to
warn you when there is an alignment issue with the struct. The semi-safe
assumption is, if the ones at the end are right then the earlier ones
are good too. You should probably still do a visual compare of a good
dump to the new dump after each server patch.